requirements.txt
fastmcp=2.3.5
pyyaml==6.0.1